Rename vid_import to vid_read
Make sure to replace "path/to/video/file.mp4" with the actual path to your video file. This code will read each frame from the video and display it in a window named "Video". It will exit the loop if the 'q' key is pressed. Finally, it releases the video object and closes the window when done.